refactor - changing ValidEnum package name to valid_enum
