+++ backend/zerowaste/src/main/java/com/zerowaste/dtos/products/EditProductDTO.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 import java.util.Set;
 
 import com.zerowaste.models.product.ProductCategory;
-import com.zerowaste.utils.validation.ValidEnum.ValidEnum;
+import com.zerowaste.utils.validation.valid_enum.ValidEnum;
 
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.Future;
